Add per workflow request object
