+++ src/DurableOrchestrator/Workflows/SplitPdfWorkflow.cs
@@ -1,7 +1,6 @@
 using DurableOrchestrator.AzureStorage;
 using DurableOrchestrator.Core;
 using DurableOrchestrator.Core.Observability;
-using DurableOrchestrator.Models;
 using iText.Kernel.Pdf;
 
 namespace DurableOrchestrator.Workflows;
@@ -23,27 +22,27 @@ public class SplitPdfWorkflow()
         [OrchestrationTrigger] TaskOrchestrationContext context)
     {
         // step 1: obtain input for the workflow
-        var workFlowInput = context.GetInput<WorkFlowInput>() ??
-                            throw new ArgumentNullException(nameof(context), $"{nameof(WorkFlowInput)} is null.");
+        var input = context.GetInput<WorkflowRequest>() ??
+                            throw new ArgumentNullException(nameof(context), $"{nameof(WorkflowRequest)} is null.");
 
-        using var span = StartActiveSpan(OrchestrationName, workFlowInput);
+        using var span = StartActiveSpan(OrchestrationName, input);
         var log = context.CreateReplaySafeLogger(OrchestrationName);
 
         var orchestrationResults = new WorkflowResult(OrchestrationName, log);
 
         // step 2: validate the input
-        var validationResult = workFlowInput.Validate();
+        var validationResult = input.Validate();
         if (!validationResult.IsValid)
         {
             orchestrationResults.AddRange(
-                nameof(IWorkflowRequest.Validate),
-                $"{nameof(workFlowInput)} is invalid.",
+                nameof(WorkflowRequest.Validate),
+                $"{nameof(input)} is invalid.",
                 validationResult.ValidationMessages,
                 LogLevel.Error);
             return orchestrationResults.Results; // Exit the orchestration due to validation errors
         }
 
-        orchestrationResults.Add(nameof(IWorkflowRequest.Validate), $"{nameof(workFlowInput)} is valid.");
+        orchestrationResults.Add(nameof(WorkflowRequest.Validate), $"{nameof(input)} is valid.");
 
         // step 3: read the source file from blob storage using the input and split the PDF file into individual pages
         var splitResults = new List<byte[]>();
@@ -51,7 +50,7 @@ public class SplitPdfWorkflow()
         var sourceFile = await CallActivityAsync<byte[]?>(
             context,
             nameof(BlobStorageActivities.GetBlobContentAsBuffer),
-            workFlowInput.SourceBlobStorageInfo!,
+            input.SourceBlobStorageInfo!,
             span.Context);
 
         if (sourceFile == null)
@@ -102,9 +101,9 @@ public class SplitPdfWorkflow()
         {
             var blobStorageRequest = new BlobStorageRequest
             {
-                StorageAccountName = workFlowInput.TargetBlobStorageInfo!.StorageAccountName,
-                ContainerName = workFlowInput.TargetBlobStorageInfo!.ContainerName,
-                BlobName = $"{workFlowInput.TargetBlobStorageInfo!.BlobName}_{i + 1}.pdf",
+                StorageAccountName = input.TargetBlobStorageInfo!.StorageAccountName,
+                ContainerName = input.TargetBlobStorageInfo!.ContainerName,
+                BlobName = $"{input.TargetBlobStorageInfo!.BlobName}_{i + 1}.pdf",
                 Buffer = splitResults[i]
             };
 
@@ -158,11 +157,76 @@ public class SplitPdfWorkflow()
         var instanceId = await StartWorkflowAsync(
             starter,
             OrchestrationName,
-            ExtractInput<WorkFlowInput>(requestBody),
+            ExtractInput<WorkflowRequest>(requestBody),
             span.Context);
 
         log.LogInformation("Started orchestration with ID = '{instanceId}'.", instanceId);
 
         return await starter.CreateCheckStatusResponseAsync(req, instanceId);
+    }
+
+    public class WorkflowRequest : IWorkflowRequest
+    {
+        [JsonPropertyName("sourceBlobStorageInfo")]
+        public BlobStorageRequest? SourceBlobStorageInfo { get; set; }
+
+        [JsonPropertyName("targetBlobStorageInfo")]
+        public BlobStorageRequest? TargetBlobStorageInfo { get; set; }
+
+        [JsonPropertyName("observableProperties")]
+        public Dictionary<string, object> ObservabilityProperties { get; set; } = new();
+
+        public ValidationResult Validate()
+        {
+            var result = new ValidationResult();
+
+            if (SourceBlobStorageInfo == null)
+            {
+                result.AddErrorMessage("Source blob storage info is missing.");
+            }
+            else
+            {
+                if (string.IsNullOrEmpty(SourceBlobStorageInfo.BlobName))
+                {
+                    // could be missing - not breaking the validity of the request
+                    result.AddMessage("Source blob name is missing.");
+                }
+
+                if (string.IsNullOrEmpty(SourceBlobStorageInfo.ContainerName))
+                {
+                    result.AddErrorMessage("Source container name is missing.");
+                }
+
+                if (string.IsNullOrEmpty(SourceBlobStorageInfo.StorageAccountName))
+                {
+                    result.AddErrorMessage("Source storage account name is missing.");
+                }
+            }
+
+            if (TargetBlobStorageInfo == null)
+            {
+                result.AddErrorMessage("Target blob storage info is missing.");
+            }
+            else
+            {
+                if (string.IsNullOrEmpty(TargetBlobStorageInfo.BlobName))
+                {
+                    // could be missing - not breaking the validity of the request
+                    result.AddMessage("Target blob name is missing.");
+                }
+
+                if (string.IsNullOrEmpty(TargetBlobStorageInfo.ContainerName))
+                {
+                    result.AddErrorMessage("Target container name is missing.");
+                }
+
+                if (string.IsNullOrEmpty(TargetBlobStorageInfo.StorageAccountName))
+                {
+                    result.AddErrorMessage("Target storage account name is missing.");
+                }
+            }
+
+            return result;
+        }
     }
 }
