+++ src/DurableOrchestrator/tests/test.rest
@@ -1,48 +1,8 @@
-
-POST http://localhost:7071/api/WorkflowOrc_HttpStart
-
-{
-  "name": "samplesec",
-  "source": "KeyVault",
-  "destination": "BlobStorage",
-  "sourceBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "functions.csv",
-    "containerName": "sourcecontainer"
-  },
-  "targetBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "whatasecret22.2.943.txt",
-    "containerName": "targetcontainer"
-  }
-}
-
-###
-POST https://durable-orc.wonderfulground-1fcf5190.uksouth.azurecontainerapps.io/api/WorkflowOrc_HttpStart
-
-{
-  "name": "samplesec",
-  "source": "sourceIdentifier",
-  "destination": "destinationIdentifier",
-  "sourceBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "asecret12.csv",
-    "containerName": "sourcecontainer"
-  },
-  "targetBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "copy_11of_functions.csv",
-    "containerName": "targetcontainer"
-  }
-}
-###
+### Run the copy blob workflow
 
 POST http://localhost:7071/api/CopyBlobWorkflow_HttpStart
 
 {
-  "name": "sampleName",
-  "source": "sourceIdentifier",
-  "destination": "destinationIdentifier",
   "sourceBlobStorageInfo": {
     "storageAccountName": "ydinventorylogs",
     "blobName": "functions.csv",
@@ -55,18 +15,63 @@ POST http://localhost:7071/api/CopyBlobWorkflow_HttpStart
   }  
 }
 
-###
+### Run the document intelligence to markdown workflow
+
+POST http://localhost:7071/api/DocumentIntelligenceToMarkdownWorkflow_HttpStart
+
+{
+  "sourceBlobStorageInfo": {
+    "storageAccountName": "ydinventorylogs",
+    "blobName": "Invoice_1.pdf",
+    "containerName": "docintelin"
+  },
+  "targetBlobStorageInfo": {
+    "storageAccountName": "ydinventorylogs",
+    "blobName": "output2.md",
+    "containerName": "docintelout"
+  }
+}
+
+### Run the key vault secret to blob workflow with split PDF sub-workflow run
+
+POST http://localhost:7071/api/KeyVaultSecretToBlobWorkflow_HttpStart
+
+{
+  "name": "samplesec",
+  "sourceBlobStorageInfo": {
+    "storageAccountName": "ydinventorylogs",
+    "blobName": "sample.pdf",
+    "containerName": "fullcontainer"
+  },
+  "targetBlobStorageInfo": {
+    "storageAccountName": "ydinventorylogs",
+    "blobName": "whatasecret22.3.943.txt",
+    "containerName": "targetcontainer"
+  }
+}
+
+### Run the Split PDF workflow
+
+POST http://localhost:7071/api/SplitPdfWorkflow_HttpStart
+
+{
+  "sourceBlobStorageInfo": {
+    "storageAccountName": "ydinventorylogs",
+    "blobName": "sample.pdf",
+    "containerName": "fullcontainer"
+  },
+  "targetBlobStorageInfo": {
+    "storageAccountName": "ydinventorylogs",
+    "blobName": "split_3.22.944_sample.pdf",
+    "containerName": "splitcontainer"
+  }  
+}
+
+### Run the semantic text analytics workflow
+
 POST http://localhost:7071/api/TextAnalyticsWorkflow_HttpStart
 
 {
-  "name": "sampleName",
-  "source": "sourceIdentifier",
-  "destination": "destinationIdentifier",
-  "sourceBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "functions.csv",
-    "containerName": "sourcecontainer"
-  },
   "targetBlobStorageInfo": {
     "storageAccountName": "ydinventorylogs",
     "blobName": "sentiments3.07.1302.json",
@@ -114,125 +119,4 @@ POST http://localhost:7071/api/TextAnalyticsWorkflow_HttpStart
       "textsToAnalyze": "Battery life has significantly worsened since the last update. It's frustrating!"
     }
   ]
-}
-
-###
-POST https://durable-orc.wonderfulground-1fcf5190.uksouth.azurecontainerapps.io/api/TextAnalyticsWorkflow_HttpStart
-
-{
-  "name": "sampleName",
-  "source": "sourceIdentifier",
-  "destination": "destinationIdentifier",
-  "sourceBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "functions.csv",
-    "containerName": "sourcecontainer"
-  },
-  "targetBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "sentiments2.26.1222.json",
-    "containerName": "targetcontainer"
-  },
-  "textAnalyticsRequests": [
-    {
-      "operationTypes": ["sentiment", "keyPhrases"],
-      "textsToAnalyze": "I absolutely love the new update! Everything is faster and more intuitive."
-    },
-    {
-      "operationTypes": ["sentiment", "keyPhrases"],
-      "textsToAnalyze": "Unfortunately, the product failed to meet my expectations. It's buggy and crashes often."
-    },
-    {
-      "operationTypes": ["sentiment", "keyPhrases"],
-      "textsToAnalyze": "The customer service was outstanding; they went above and beyond to solve my issue."
-    },
-    {
-      "operationTypes": ["sentiment", "keyPhrases"],
-      "textsToAnalyze": "I'm indifferent about the changes. I liked some improvements, but I think others were unnecessary."
-    },
-    {
-      "operationTypes": ["sentiment", "keyPhrases"],
-      "textsToAnalyze": "The installation process was a nightmare. Instructions were not clear, and it took hours to set up."
-    },
-    {
-      "operationTypes": ["sentiment", "keyPhrases"],
-      "textsToAnalyze": "This is by far the best version of the product. It's sleek, efficient, and a joy to use."
-    },
-    {
-      "operationTypes": ["sentiment", "keyPhrases"],
-      "textsToAnalyze": "The product feels outdated and lacks innovative features compared to competitors."
-    },
-    {
-      "operationTypes": ["sentiment", "keyPhrases"],
-      "textsToAnalyze": "I had high hopes, but the app constantly freezes on my phone. Very disappointed."
-    },
-    {
-      "operationTypes": ["sentiment", "keyPhrases"],
-      "textsToAnalyze": "The new design is fantastic! It's clean, modern, and easy to navigate."
-    },
-    {
-      "operationTypes": ["sentiment", "keyPhrases"],
-      "textsToAnalyze": "Battery life has significantly worsened since the last update. It's frustrating!"
-    }
-  ]
-}
-
-### Split PDF
-
-POST http://localhost:7071/api/SplitPdfWorkflow_HttpStart
-
-{
-  "name": "SplitPDF",
-  "source": "sourceIdentifier",
-  "destination": "destinationIdentifier",
-  "sourceBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "sample.pdf",
-    "containerName": "fullcontainer"
-  },
-  "targetBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "split_3.22.944_sample.pdf",
-    "containerName": "splitcontainer"
-  }  
-}
-
-### calling a workflow from a workflow
-
-POST http://localhost:7071/api/WorkflowOrc_HttpStart
-
-{
-  "name": "samplesec",
-  "source": "KeyVault",
-  "destination": "BlobStorage",
-  "sourceBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "sample.pdf",
-    "containerName": "fullcontainer"
-  },
-  "targetBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "whatasecret22.3.943.txt",
-    "containerName": "targetcontainer"
-  }
-}
-
-### calling DI
-
-POST http://localhost:7071/api/DILayout2mdWorkFlow_HttpStart
-
-{
-  "name": "samplesec",
-  "source": "KeyVault",
-  "destination": "BlobStorage",
-  "sourceBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "Invoice_1.pdf",
-    "containerName": "docintelin"
-  },
-  "targetBlobStorageInfo": {
-    "storageAccountName": "ydinventorylogs",
-    "blobName": "output2.md",
-    "containerName": "docintelout"
-  }
 }
